VBoxHeadless: warning.
git-svn-id: http://www.virtualbox.org/svn/vbox/trunk@49980 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Frontends/VBoxHeadless/Framebuffer.cpp
@@ -44,7 +44,7 @@ VRDPFramebuffer::VRDPFramebuffer()
 
     // start with a standard size
     RequestResize(0, FramebufferPixelFormat_Opaque,
-                  (ULONG) NULL, 0, 0, 640, 480, NULL);
+                  NULL, 0, 0, 640, 480, NULL);
 }
 
 VRDPFramebuffer::~VRDPFramebuffer()
